minor refactoring

+++ commons-components/src/main/java/io/github/gustavosdelgado/library/service/AuthTokenService.java
@@ -1,22 +1,20 @@
 package io.github.gustavosdelgado.library.service;
 
-import java.time.Clock;
-import java.time.Instant;
-import java.time.ZoneId;
-
-import io.github.gustavosdelgado.library.exception.NoDataFoundException;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.JWTVerifier.BaseVerification;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.interfaces.DecodedJWT;
-
 import io.github.gustavosdelgado.library.domain.user.User;
+import io.github.gustavosdelgado.library.exception.NoDataFoundException;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
 
 @Service
 public class AuthTokenService {
@@ -25,7 +23,7 @@ public class AuthTokenService {
     private String secret;
 
     @Value("${api.security.token.expiration}")
-    private int expirationInMinutes;
+    private long expirationInMinutes;
 
     @Value("${api.security.token.timezone}")
     private String tokenTimeZone;
